# api/openapi.yaml
openapi: 3.1.0
info:
  title: 神州 HR 考勤系统 API
  version: 1.9.0-wave3
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: |
    V1.9 WAVE-3 契约：在完整保留 WAVE-1 本地账号、登录会话、权限、审计与
    版本化规则配置底座，以及 WAVE-2 组织、员工期初导入与本地人员权威的基础上，
    增加考勤地点、考勤组、员工期限分配、不可变班次版本、工作日历和基础策略绑定。
    所有 ID 均按 string 传输；响应和示例不包含密码、密码哈希、session token 或 reset token。
servers:
  - url: /api/v1
security:
  - sessionCookie: []
...
paths:
...
  /organization-units:
    get:
      tags: [Compatibility]
      operationId: getCurrentOrganizationTree
      summary: 获取当前主体授权范围内的组织当前投影
      x-capability: MASTER_DATA:READ
      x-data-scope: PEOPLE:ORGANIZATION
      parameters:
        - name: asOf
          in: query
          schema: { type: string, format: date }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - name: includeInactive
          in: query
          schema: { type: boolean, default: false }
      responses:
        '200':
          description: 组织树
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrganizationNode'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '404':
          $ref: '#/components/responses/ResourceNotAvailable'
    post:
      tags: [People]
      operationId: createLocalOrganization
      summary: 创建本地权威组织及首个不可变版本
      x-capability: ORGANIZATION:CREATE
      x-data-scope: PEOPLE:LEGAL_ENTITY
      x-audit-event: ORGANIZATION_VERSION_CREATED
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationCreateRequest'
      responses:
        '201':
          description: 组织及首个本地版本
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationDetail'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/AccessDenied'
        '409':
          $ref: '#/components/responses/PeopleConflict'
  /organization-units/{organizationId}:
    get:
      tags: [People]
      operationId: getOrganization
      summary: 查询本地权威组织详情
      x-capability: ORGANIZATION:READ
      x-data-scope: PEOPLE:ORGANIZATION
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
        - name: asOf
          in: query
          schema: { type: string, format: date }
      responses:
        '200':
          description: 组织详情
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationDetail'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/AccessDenied'
        '404':
          $ref: '#/components/responses/ResourceNotAvailable'
    patch:
      tags: [People]
      operationId: updateLocalOrganization
      summary: 以新版本更新本地组织，不覆盖来源文件或发布快照
      x-capability: ORGANIZATION:EDIT
      x-data-scope: PEOPLE:ORGANIZATION
      x-audit-event: ORGANIZATION_VERSION_CREATED
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationUpdateRequest'
      responses:
        '200':
          description: 新的当前组织版本
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationDetail'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/AccessDenied'
        '404':
          $ref: '#/components/responses/ResourceNotAvailable'
        '409':
          $ref: '#/components/responses/PeopleConflict'
  /organization-units/{organizationId}/versions:
    get:
      tags: [People]
      operationId: listOrganizationVersions
      summary: 查询组织不可变版本历史
      x-capability: ORGANIZATION:READ
      x-data-scope: PEOPLE:ORGANIZATION
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: 组织版本历史
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationVersionPage'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/AccessDenied'
        '404':
          $ref: '#/components/responses/ResourceNotAvailable'
  /employees:
    get:
      tags: [Compatibility]
      operationId: listEmployees
      summary: 分页获取当前主体授权范围内的人员摘要
      x-capability: MASTER_DATA:READ
      x-data-scope: PEOPLE:ORGANIZATION
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - name: query
          in: query
          schema: { type: string, maxLength: 100 }
        - name: organizationId
          in: query
          schema: { $ref: '#/components/schemas/ResourceId' }
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/EmployeeStatus' }
        - name: asOf
          in: query
          schema: { type: string, format: date }
        - name: sort
          in: query
          schema:
            type: string
            enum: [employeeNumber, displayName, employmentStatus, organizationName, updatedAt]
      responses:
        '200':
          description: 授权范围内人员摘要；不含证件、报酬或精确位置字段
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeePage'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '404':
          $ref: '#/components/responses/ResourceNotAvailable'
    post:
      tags: [People]
      operationId: createLocalEmployee
      summary: 创建本地员工及首个不可变版本
      x-capability: EMPLOYEE:CREATE
      x-data-scope: PEOPLE:ORGANIZATION
      x-audit-event: EMPLOYEE_VERSION_CREATED
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeCreateRequest'
      responses:
        '201':
          description: 员工及首个本地版本
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeDetail'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/AccessDenied'
        '409':
          $ref: '#/components/responses/PeopleConflict'
  /employees/{employeeId}:
    get:
      tags: [People]
      operationId: getEmployee
      summary: 查询员工、当前任职和累计工龄详情
      x-capability: EMPLOYEE:READ
      x-data-scope: PEOPLE:ORGANIZATION
      parameters:
        - $ref: '#/components/parameters/EmployeeId'
        - name: asOf
          in: query
          schema: { type: string, format: date }
      responses:
        '200':
          description: 员工详情
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeDetail'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/AccessDenied'
        '404':
          $ref: '#/components/responses/ResourceNotAvailable'
    patch:
      tags: [People]
      operationId: updateLocalEmployee
      summary: 以新版本更新员工，不覆盖来源行或发布快照
      x-capability: EMPLOYEE:EDIT
      x-data-scope: PEOPLE:ORGANIZATION
      x-audit-event: EMPLOYEE_VERSION_CREATED
      parameters:
        - $ref: '#/components/parameters/EmployeeId'
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeUpdateRequest'
      responses:
        '200':
          description: 新的当前员工版本
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeDetail'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/AccessDenied'
        '404':
          $ref: '#/components/responses/ResourceNotAvailable'
        '409':
          $ref: '#/components/responses/PeopleConflict'
  /employees/{employeeId}/versions:
    get:
      tags: [People]
      operationId: listEmployeeVersions
      summary: 查询员工不可变版本历史
      x-capability: EMPLOYEE:READ
      x-data-scope: PEOPLE:ORGANIZATION
      parameters:
        - $ref: '#/components/parameters/EmployeeId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: 员工版本历史
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeVersionPage'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/AccessDenied'
        '404':
          $ref: '#/components/responses/ResourceNotAvailable'
  /employees/{employeeId}/employment-periods:
    get:
      tags: [People]
      operationId: listEmploymentPeriods
      summary: 按历史日期查询全部任职周期
      x-capability: EMPLOYMENT:READ
      x-data-scope: PEOPLE:ORGANIZATION
      parameters:
        - $ref: '#/components/parameters/EmployeeId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - name: asOf
          in: query
          schema: { type: string, format: date }
      responses:
        '200':
          description: 半开区间任职历史；空档日期返回空 items，不能猜测归属
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmploymentPeriodPage'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/AccessDenied'
        '404':
          $ref: '#/components/responses/ResourceNotAvailable'
    post:
      tags: [People]
      operationId: createEmploymentPeriod
      summary: 创建新的任职周期；二次入职不得覆盖历史周期
      x-capability: EMPLOYMENT:CREATE
      x-data-scope: PEOPLE:ORGANIZATION
      x-audit-event: EMPLOYMENT_PERIOD_CREATED
      parameters:
        - $ref: '#/components/parameters/EmployeeId'
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmploymentPeriodCreateRequest'
      responses:
        '201':
          description: 新任职周期
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmploymentPeriodView'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/AccessDenied'
        '404':
          $ref: '#/components/responses/ResourceNotAvailable'
        '409':
          $ref: '#/components/responses/PeopleConflict'
  /employees/{employeeId}/employment-periods/{employmentPeriodId}:
    patch:
      tags: [People]
      operationId: updateEmploymentPeriod
      summary: 以前向新版本更正任职或登记离职
      description: 业务 terminationDate 在服务端转换为 `endExclusive=terminationDate+1 day`。
      x-capability: EMPLOYMENT:EDIT
      x-data-scope: PEOPLE:ORGANIZATION
      x-audit-event: EMPLOYMENT_PERIOD_VERSION_CREATED
      parameters:
        - $ref: '#/components/parameters/EmployeeId'
        - $ref: '#/components/parameters/EmploymentPeriodId'
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmploymentPeriodUpdateRequest'
      responses:
        '200':
          description: 新的任职周期版本
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmploymentPeriodView'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/AccessDenied'
        '404':
          $ref: '#/components/responses/ResourceNotAvailable'
        '409':
          $ref: '#/components/responses/PeopleConflict'
  /employees/{employeeId}/prior-service-records:
    get:
      tags: [People]
      operationId: listPriorServiceRecords
      summary: 查询可重放的入职前累计工龄发生额和调整记录
      x-capability: PRIOR_SERVICE:READ
      x-data-scope: PEOPLE:ORGANIZATION
      parameters:
        - $ref: '#/components/parameters/EmployeeId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: 入职前累计工龄记录及重放合计
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PriorServiceRecordPage'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/AccessDenied'
        '404':
          $ref: '#/components/responses/ResourceNotAvailable'
  /employees/{employeeId}/prior-service-adjustments:
    post:
      tags: [People]
      operationId: createPriorServiceAdjustment
      summary: 追加累计工龄调整记录，不覆盖历史发生额
      x-capability: PRIOR_SERVICE:ADJUST
      x-data-scope: PEOPLE:ORGANIZATION
      x-audit-event: PRIOR_SERVICE_ADJUSTED
      parameters:
        - $ref: '#/components/parameters/EmployeeId'
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PriorServiceAdjustmentRequest'
      responses:
        '201':
          description: 新增发生额及重放后的累计结果
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PriorServiceRecordView'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/AccessDenied'
        '404':
          $ref: '#/components/responses/ResourceNotAvailable'
        '409':
          $ref: '#/components/responses/PeopleConflict'
  /employees/{employeeId}/prior-service/recalculate:
    post:
      tags: [People]
      operationId: recalculatePriorService
      summary: 重放全部发生额并返回可复现的累计工龄
      x-capability: PRIOR_SERVICE:ADJUST
      x-data-scope: PEOPLE:ORGANIZATION
      x-audit-event: PRIOR_SERVICE_RECALCULATED
      parameters:
        - $ref: '#/components/parameters/EmployeeId'
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReasonRequest'
      responses:
        '200':
          description: 确定性重放结果
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PriorServiceReplayView'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/AccessDenied'
        '404':
          $ref: '#/components/responses/ResourceNotAvailable'
        '409':
          $ref: '#/components/responses/PeopleConflict'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: SHENZHOUHR_SESSION
      description: 本系统本地账号安全会话；开发身份头不属于正式认证路径。
...
  parameters:
...
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
    Size:
      name: size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
  schemas:
...
    OrganizationNode:
      type: object
      additionalProperties: false
      required:
        [organizationId, organizationVersionId, code, name, organizationType,
         status, effectiveFrom, sourceAuthority, rowVersion, children]
      properties:
        organizationId: { $ref: '#/components/schemas/ResourceId' }
        organizationVersionId: { $ref: '#/components/schemas/ResourceId' }
        code: { type: string }
        name: { type: string }
        organizationType: { type: string, enum: [COMPANY, DEPARTMENT, TEAM] }
        status: { $ref: '#/components/schemas/OrganizationStatus' }
        sourceOrganizationId:
          oneOf:
            - $ref: '#/components/schemas/ExternalPreciseId'
            - type: 'null'
        effectiveFrom: { type: string, format: date }
        effectiveTo: { type: [string, 'null'], format: date }
        sourceAuthority: { $ref: '#/components/schemas/PeopleSourceAuthority' }
        rowVersion: { type: integer, format: int64, minimum: 0 }
        children:
          type: array
          items:
            $ref: '#/components/schemas/OrganizationNode'
...
  responses:
    AuthenticationRequired:
      description: 身份无效、会话到期或会话已撤销
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...